+++ jschon/catalog/_2019_09.py
@@ -13,7 +13,7 @@ from jschon.vocabulary.validation import *
 def initialize(catalog: Catalog):
     catalog.add_uri_source(
         URI('https://json-schema.org/draft/2019-09/'),
-        LocalSource(pathlib.Path(__file__).parent / 'json-schema-spec-2019-09', suffix='.json'),
+        LocalSource(pathlib.Path(__file__).parent / 'json-schema-2019-09', suffix='.json'),
     )
 
     catalog.create_vocabulary(