+++ CHANGELOG.rst
@@ -3,6 +3,11 @@ Changelog
 
 v0.10.1 (in development)
 ------------------------
+Miscellaneous:
+
+* Rename submodules `json-schema-spec-2019-09` and `json-schema-spec-2020-12` to
+  `json-schema-2019-09` and `json-schema-2020-12`, respectively
+  (run ``git submodule init`` to update local git config)
 
 
 v0.10.0 (2023-04-01)